Embedded computing lab task 4
In this we havew to create 2 python program
1. to control motor and read the encoder
2. To read the linefollower sensor and have a TCP and
UDS sever to get sensor data from C program and send
collected data to webserver.

This also contains the c program which will be reading the SR04 data
every 50ms and send the data through the TCP socket.

In the webpage we will be ablle to start the task by checking the
checkbox. it also displays the data received.